* updating css branch from trunk

+++ branches/CSSBranch/examples/advanced/physics/physicsShapes/PhysicsCircle.java
@@ -29,7 +29,7 @@ public class PhysicsCircle extends MTEllipse implements IPhysicsComponent{
 			World world, float density, float friction, float restitution, float worldScale
 	) {
 		//super(applet, centerPoint, radius/(float)worldScale, radius/(float)worldScale);
-		super(applet, centerPoint, radius/(float)worldScale, radius/(float)worldScale);
+		super(applet, centerPoint, radius/ worldScale, radius/ worldScale);
 		this.angle = 0;
 		this.world = world;
 		this.density = density;
@@ -40,7 +40,7 @@ public class PhysicsCircle extends MTEllipse implements IPhysicsComponent{
 		this.setGestureAllowance(RotateProcessor.class, false);
 		
 		BodyDef dymBodyDef = new BodyDef();
-		dymBodyDef.position = new Vec2(centerPoint.x /(float)worldScale, centerPoint.y /(float)worldScale);
+		dymBodyDef.position = new Vec2(centerPoint.x / worldScale, centerPoint.y / worldScale);
 		this.bodyDefB4CreationCallback(dymBodyDef);
 		this.body = this.world.createBody(dymBodyDef);
 		
@@ -48,7 +48,7 @@ public class PhysicsCircle extends MTEllipse implements IPhysicsComponent{
 //		circleDef.radius = radius/(float)worldScale; 
 		 //FIXME HACK so textured circles really connect to other bodies
 //		circleDef.radius = radius/(float)worldScale - 2/(float)worldScale; 
-		circleDef.radius = radius/(float)worldScale;
+		circleDef.radius = radius/ worldScale;
 		if (density != 0.0f){
 			circleDef.density 		= density;
 			circleDef.friction 		= friction;
